+++ src/app.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <array>
+#include <cctype>
 #include <cstring>
 #include <ctime>
 #include <inkview.h>
@@ -8,6 +9,7 @@
 #include <memory>
 #include <string>
 #include <type_traits>
+#include <vector>
 
 #include "config.h"
 #include "errors.h"
@@ -16,6 +18,7 @@
 #include "model.h"
 #include "service.h"
 #include "ui.h"
+#include "util.h"
 
 #define T(x) GetLangText(x)
 
@@ -51,10 +54,8 @@ public:
     }
 
     if (event_type == EVT_CONFIGCHANGED) {
-      this->load_config();
-      if (this->ui) {
-        this->ui->show();
-      }
+      this->clear_model_weather_conditions();
+      this->show();
       return 1;
     }
 
@@ -63,12 +64,13 @@ public:
     }
 
     if (this->ui) {
-      if (event_type == EVT_POINTERDOWN || event_type == EVT_POINTERDRAG ||
-          event_type == EVT_POINTERUP) {
+      if (ISPOINTEREVENT(event_type)) {
         return this->ui->handle_pointer_event(event_type, param_one, param_two);
       }
-      if (event_type == EVT_KEYPRESS) {
-        return this->ui->handle_key_pressed(param_one);
+      if (ISKEYEVENT(event_type)) {
+        if (event_type == EVT_KEYPRESS) {
+          return this->ui->handle_key_pressed(param_one);
+        }
       }
     }
 
@@ -90,7 +92,7 @@ private:
 
     this->load_config();
     this->ui->show();
-    this->refresh_model();
+    this->refresh_model_weather_conditions();
   }
 
   void exit() {
@@ -119,25 +121,46 @@ private:
   }
 
   int handle_custom_event(int param_one, int param_two) {
-    if (param_one == CustomEvent::model_updated) {
+    if (param_one == CustomEvent::config_editor_requested) {
+      this->open_config_editor();
+      return 1;
+    } else if (param_one == CustomEvent::about_dialog_requested) {
+      this->open_about_dialog();
+      return 1;
+    } else if (param_one == CustomEvent::model_updated) {
       if (this->ui) {
         this->ui->show();
         return 1;
       }
+    } else if (param_one == CustomEvent::new_location_requested) {
+      auto *const raw_location =
+          reinterpret_cast<std::array<char, 256> *>(GetCurrentEventExData());
+      const std::string location{raw_location->data()};
+      update_config_location(location);
+      return 1;
     } else if (param_one == CustomEvent::refresh_requested) {
-      this->refresh_model();
+      this->refresh_model_weather_conditions();
       return 1;
-
-    } else if (param_one == CustomEvent::config_editor_requested) {
-      this->open_config_editor();
-    } else if (param_one == CustomEvent::about_dialog_requested) {
-      this->open_about_dialog();
-      return 1;
+    } else if (param_one == CustomEvent::warning_emitted) {
+      if (param_two == CustomEventParam::invalid_location) {
+        DialogSynchro(
+            ICON_WARNING, T("Invalid input"),
+            T("Make sure you enter a town then a country separated by a "
+              "comma."),
+            T("Ok"), nullptr, nullptr);
+        return 1;
+      }
     }
     return 0;
   }
 
-  void refresh_model() {
+  void clear_model_weather_conditions() {
+    this->model->current_condition = std::optional<Condition>{};
+    this->model->hourly_forecast.clear();
+    this->model->refresh_date = 0;
+  }
+
+  void refresh_model_weather_conditions() {
     ShowHourglassForce();
 
     this->model->refresh_date = std::time(nullptr);
@@ -186,6 +209,50 @@ private:
     const auto about_content = get_about_content();
     Dialog(ICON_INFORMATION, T("About"), about_content.c_str(), T("Ok"),
            nullptr, &handle_about_dialog_button_clicked);
+  }
+
+  static std::pair<std::string, std::string>
+  extract_town_and_country(const std::string &location) {
+    std::stringstream to_parse{location};
+    std::vector<std::string> tokens;
+    std::string token;
+    while (std::getline(to_parse, token, ',')) {
+      ltrim(token);
+      rtrim(token);
+      tokens.push_back(token);
+    }
+    if (tokens.size() == 1) {
+      auto town = tokens[0];
+      town[0] = std::toupper(town[0]);
+      return {tokens[0], ""};
+    } else if (tokens.size() > 1) {
+      auto country = tokens[tokens.size() - 1];
+      country[0] = std::toupper(country[0]);
+
+      auto town = location.substr(0, location.size() - country.size());
+      ltrim(town);
+      rtrim(town);
+      town[0] = std::toupper(town[0]);
+
+      return {town, country};
+    }
+    throw InvalidLocation{};
+  }
+
+  void update_config_location(const std::string &location) {
+    try {
+      auto [town, country] = App::extract_town_and_country(location);
+
+      Config config;
+      config.write_string("location_town", town);
+      config.write_string("location_country", country);
+
+      Config::config_changed();
+    } catch (const InvalidLocation &error) {
+      const auto event_handler = GetEventHandler();
+      SendEvent(event_handler, EVT_CUSTOM, CustomEvent::warning_emitted,
+                CustomEventParam::invalid_location);
+    }
   }
 };