+++ src/events.h
@@ -1,8 +1,14 @@
 #pragma once
 
 enum CustomEvent {
-  model_updated,
-  refresh_requested,
+  about_dialog_requested,
   config_editor_requested,
-  about_dialog_requested
+  model_updated,
+  new_location_requested,
+  refresh_requested,
+  warning_emitted,
+};
+
+enum CustomEventParam {
+  invalid_location,
 };